Move admin/rest_api to actix_web_1 crate
This commit moves the admin/rest_api code from libsplinter into the
actix_web_1 crate and updates all the internal use statements.

It also updates the docker files to copy the actix-web-1 and common
crates into the build directory in our docker images.
