+++ libsplinter/src/admin/mod.rs
@@ -19,8 +19,6 @@ pub mod client;
 pub mod error;
 pub mod lifecycle;
 pub mod messages;
-#[cfg(any(feature = "rest-api-actix-web-1", feature = "rest-api-actix-web-3"))]
-pub mod rest_api;
 pub mod service;
 pub mod store;
 mod token;
