+++ libsplinter/src/rest_api/mod.rs
@@ -101,7 +101,6 @@ pub use actix_web_1::{
 };
 
 #[cfg(any(
-    feature = "admin-service",
     feature = "authorization",
     feature = "biome-credentials",
     feature = "biome-key-management",
