+++ rest_api/actix_web_1/src/admin/circuits.rs
@@ -19,17 +19,18 @@ use actix_web::{error::BlockingError, web, Error, HttpRequest, HttpResponse};
 use futures::{future::IntoFuture, Future};
 use std::collections::HashMap;
 
-#[cfg(feature = "authorization")]
-use crate::admin::rest_api::CIRCUIT_READ_PERMISSION;
-use crate::admin::store::{AdminServiceStore, CircuitPredicate, CircuitStatus};
-use crate::rest_api::{
+use splinter::admin::store::{AdminServiceStore, CircuitPredicate, CircuitStatus};
+use splinter::rest_api::{
     actix_web_1::{Method, ProtocolVersionRangeGuard, Resource},
     paging::{get_response_paging_info, DEFAULT_LIMIT, DEFAULT_OFFSET},
-    ErrorResponse, SPLINTER_PROTOCOL_VERSION,
+    ErrorResponse,
 };
+use splinter_rest_api_common::SPLINTER_PROTOCOL_VERSION;
 
-use super::super::error::CircuitListError;
-use super::super::resources;
+use super::error::CircuitListError;
+use super::resources;
+#[cfg(feature = "authorization")]
+use super::CIRCUIT_READ_PERMISSION;
 
 const ADMIN_LIST_CIRCUITS_MIN: u32 = 1;
 
@@ -248,18 +249,20 @@ mod tests {
     use reqwest::{blocking::Client, StatusCode, Url};
     use serde_json::{to_value, Value as JsonValue};
 
-    use crate::admin::store::diesel::DieselAdminServiceStore;
-    use crate::admin::store::{
+    use splinter::admin::store::diesel::DieselAdminServiceStore;
+    use splinter::admin::store::{
         AuthorizationType, Circuit, CircuitBuilder, CircuitNode, CircuitNodeBuilder,
         DurabilityType, PersistenceType, RouteType, ServiceBuilder,
     };
-    use crate::error::InternalError;
-    use crate::migrations::run_sqlite_migrations;
-    use crate::rest_api::actix_web_1::AuthConfig;
-    use crate::rest_api::auth::authorization::{AuthorizationHandler, AuthorizationHandlerResult};
-    use crate::rest_api::auth::identity::{Identity, IdentityProvider};
-    use crate::rest_api::auth::AuthorizationHeader;
-    use crate::rest_api::{
+    use splinter::error::InternalError;
+    use splinter::migrations::run_sqlite_migrations;
+    use splinter::rest_api::actix_web_1::AuthConfig;
+    use splinter::rest_api::auth::authorization::{
+        AuthorizationHandler, AuthorizationHandlerResult,
+    };
+    use splinter::rest_api::auth::identity::{Identity, IdentityProvider};
+    use splinter::rest_api::auth::AuthorizationHeader;
+    use splinter::rest_api::{
         actix_web_1::{RestApiBuilder, RestApiShutdownHandle},
         paging::Paging,
     };
@@ -814,7 +817,7 @@ mod tests {
         #[cfg(not(feature = "https-bind"))]
         let bind = "127.0.0.1:0";
         #[cfg(feature = "https-bind")]
-        let bind = crate::rest_api::BindConfig::Http("127.0.0.1:0".into());
+        let bind = splinter::rest_api::BindConfig::Http("127.0.0.1:0".into());
 
         let identity_provider = MockIdentityProvider::default().clone_box();
         let auth_config = AuthConfig::Custom {
