+++ rest_api/actix_web_1/src/admin/circuits_circuit_id.rs
@@ -18,16 +18,17 @@
 use actix_web::{error::BlockingError, web, Error, HttpRequest, HttpResponse};
 use futures::Future;
 
-#[cfg(feature = "authorization")]
-use crate::admin::rest_api::CIRCUIT_READ_PERMISSION;
-use crate::admin::store::AdminServiceStore;
-use crate::rest_api::{
+use splinter::admin::store::AdminServiceStore;
+use splinter::rest_api::{
     actix_web_1::{Method, ProtocolVersionRangeGuard, Resource},
-    ErrorResponse, SPLINTER_PROTOCOL_VERSION,
+    ErrorResponse,
 };
+use splinter_rest_api_common::SPLINTER_PROTOCOL_VERSION;
 
-use super::super::error::CircuitFetchError;
-use super::super::resources;
+use super::error::CircuitFetchError;
+use super::resources;
+#[cfg(feature = "authorization")]
+use super::CIRCUIT_READ_PERMISSION;
 
 const ADMIN_FETCH_CIRCUIT_MIN: u32 = 1;
 
@@ -131,18 +132,20 @@ mod tests {
     use reqwest::{blocking::Client, StatusCode, Url};
     use serde_json::{to_value, Value as JsonValue};
 
-    use crate::admin::store::diesel::DieselAdminServiceStore;
-    use crate::admin::store::{
+    use splinter::admin::store::diesel::DieselAdminServiceStore;
+    use splinter::admin::store::{
         AuthorizationType, Circuit, CircuitBuilder, CircuitNode, CircuitNodeBuilder,
         DurabilityType, PersistenceType, RouteType, ServiceBuilder,
     };
-    use crate::error::InternalError;
-    use crate::migrations::run_sqlite_migrations;
-    use crate::rest_api::actix_web_1::AuthConfig;
-    use crate::rest_api::actix_web_1::{RestApiBuilder, RestApiShutdownHandle};
-    use crate::rest_api::auth::authorization::{AuthorizationHandler, AuthorizationHandlerResult};
-    use crate::rest_api::auth::identity::{Identity, IdentityProvider};
-    use crate::rest_api::auth::AuthorizationHeader;
+    use splinter::error::InternalError;
+    use splinter::migrations::run_sqlite_migrations;
+    use splinter::rest_api::actix_web_1::AuthConfig;
+    use splinter::rest_api::actix_web_1::{RestApiBuilder, RestApiShutdownHandle};
+    use splinter::rest_api::auth::authorization::{
+        AuthorizationHandler, AuthorizationHandlerResult,
+    };
+    use splinter::rest_api::auth::identity::{Identity, IdentityProvider};
+    use splinter::rest_api::auth::AuthorizationHeader;
 
     #[test]
     /// Tests a GET /admin/circuit/{circuit_id} request returns the expected circuit.
@@ -350,7 +353,7 @@ mod tests {
         #[cfg(not(feature = "https-bind"))]
         let bind = "127.0.0.1:0";
         #[cfg(feature = "https-bind")]
-        let bind = crate::rest_api::BindConfig::Http("127.0.0.1:0".into());
+        let bind = splinter::rest_api::BindConfig::Http("127.0.0.1:0".into());
         let identity_provider = MockIdentityProvider::default().clone_box();
         let auth_config = AuthConfig::Custom {
             resources: Vec::new(),
