+++ rest_api/actix_web_1/src/admin/proposals_circuit_id.rs
@@ -20,16 +20,17 @@ use std::convert::TryFrom;
 use actix_web::{error::BlockingError, web, Error, HttpRequest, HttpResponse};
 use futures::Future;
 
-use crate::admin::rest_api::error::ProposalFetchError;
-#[cfg(feature = "authorization")]
-use crate::admin::rest_api::CIRCUIT_READ_PERMISSION;
-use crate::admin::service::proposal_store::ProposalStoreFactory;
-use crate::rest_api::{
+use splinter::admin::service::proposal_store::ProposalStoreFactory;
+use splinter::rest_api::{
     actix_web_1::{Method, ProtocolVersionRangeGuard, Resource},
-    ErrorResponse, SPLINTER_PROTOCOL_VERSION,
+    ErrorResponse,
 };
+use splinter_rest_api_common::SPLINTER_PROTOCOL_VERSION;
 
-use super::super::resources;
+use super::error::ProposalFetchError;
+use super::resources;
+#[cfg(feature = "authorization")]
+use super::CIRCUIT_READ_PERMISSION;
 
 const ADMIN_FETCH_PROPOSALS_PROTOCOL_MIN: u32 = 1;
 
@@ -143,7 +144,7 @@ mod tests {
     use reqwest::{blocking::Client, StatusCode, Url};
     use serde_json::{to_value, Value as JsonValue};
 
-    use crate::admin::{
+    use splinter::admin::{
         messages::{
             AuthorizationType, CircuitProposal, CircuitStatus, CreateCircuit, DurabilityType,
             PersistenceType, ProposalType, RouteType,
@@ -153,12 +154,14 @@ mod tests {
         },
         store::CircuitPredicate,
     };
-    use crate::error::InternalError;
-    use crate::rest_api::actix_web_1::AuthConfig;
-    use crate::rest_api::actix_web_1::{RestApiBuilder, RestApiShutdownHandle};
-    use crate::rest_api::auth::authorization::{AuthorizationHandler, AuthorizationHandlerResult};
-    use crate::rest_api::auth::identity::{Identity, IdentityProvider};
-    use crate::rest_api::auth::AuthorizationHeader;
+    use splinter::error::InternalError;
+    use splinter::rest_api::actix_web_1::AuthConfig;
+    use splinter::rest_api::actix_web_1::{RestApiBuilder, RestApiShutdownHandle};
+    use splinter::rest_api::auth::authorization::{
+        AuthorizationHandler, AuthorizationHandlerResult,
+    };
+    use splinter::rest_api::auth::identity::{Identity, IdentityProvider};
+    use splinter::rest_api::auth::AuthorizationHeader;
 
     #[test]
     /// Tests a GET /admin/proposals/{circuit_id} request returns the expected proposal.
@@ -322,7 +325,7 @@ mod tests {
         #[cfg(not(feature = "https-bind"))]
         let bind = "127.0.0.1:0";
         #[cfg(feature = "https-bind")]
-        let bind = crate::rest_api::BindConfig::Http("127.0.0.1:0".into());
+        let bind = splinter::rest_api::BindConfig::Http("127.0.0.1:0".into());
         let identity_provider = MockIdentityProvider::default().clone_box();
         let auth_config = AuthConfig::Custom {
             resources: Vec::new(),
