+++ rest_api/actix_web_1/src/admin/resources/mod.rs
