+++ rest_api/actix_web_1/src/admin/resources/v2/circuits_circuit_id.rs
@@ -14,15 +14,9 @@
 
 use std::collections::BTreeMap;
 
-use crate::admin::store::{Circuit, CircuitNode, CircuitStatus, Service};
-use crate::hex::to_hex;
-use crate::rest_api::paging::Paging;
+use splinter::admin::store::{Circuit, CircuitNode, CircuitStatus, Service};
 
-#[derive(Debug, Serialize, Clone, PartialEq)]
-pub(crate) struct ListCircuitsResponse<'a> {
-    pub data: Vec<CircuitResponse<'a>>,
-    pub paging: Paging,
-}
+use crate::hex::to_hex;
 
 #[derive(Debug, Serialize, Clone, PartialEq)]
 pub(crate) struct CircuitResponse<'a> {
