+++ tests/Dockerfile
@@ -24,9 +24,9 @@ COPY README.md /build/README.md
 COPY libsplinter /build/libsplinter
 COPY splinterd /build/splinterd
 COPY cli /build/cli
-COPY rest_api/actix_web_1/Cargo.toml /build/rest_api/actix_web_1/Cargo.toml
+COPY rest_api/actix_web_1/ /build/rest_api/actix_web_1/
 COPY rest_api/actix_web_4/Cargo.toml /build/rest_api/actix_web_4/Cargo.toml
-COPY rest_api/common/Cargo.toml /build/rest_api/common/Cargo.toml
+COPY rest_api/common/ /build/rest_api/common/
 COPY services/scabbard/cli /build/services/scabbard/cli
 COPY services/scabbard/libscabbard /build/services/scabbard/libscabbard
 COPY examples/gameroom/database /build/examples/gameroom/database
